Move model classes around to match guidelines
This also moves some other files around for consistency.
