+++ sdk/core/AzureCore/Source/Pipeline/Policies/HeadersValidationPolicy.swift
@@ -62,10 +62,9 @@ public class HeadersValidationPolicy: PipelineStage {
                     requestValue = "REDACTED"
                     responseValue = "REDACTED"
                 }
-                error = AzureError
-                    .general(
-                        "Value for header '\(key)' did not match. Expected: \(requestValue ?? "nil") Actual: \(responseValue)"
-                    )
+                error = AzureError.general(
+                    "Value for header '\(key)' did not match. Expected: \(requestValue ?? "nil") Actual: \(responseValue)"
+                )
             }
         }
         completionHandler(pipelineResponse, error)
