+++ sdk/storage/AzureStorageBlob/Source/DataStructures/NetworkingPolicy.swift
