+++ sdk/storage/AzureStorageBlob/Source/Models/Enumerations.swift
@@ -85,79 +85,3 @@ public enum LeaseStatus: String, Codable {
     /// The lease is unlocked.
     case unlocked
 }
-
-/// The state of a transfer.
-public enum TransferState: Int16 {
-    /// The transfer has not been started by the transfer management engine.
-    case pending
-    /// The transfer is currently in progress.
-    case inProgress
-    /// The transfer is paused.
-    case paused
-    /// The transfer completed successfully.
-    case complete
-    /// The transfer failed. This failure may or may not be retryable.
-    case failed
-    /// The transfer was explicitly canceled.
-    case canceled
-    /// The record of the transfer no longer exists.
-    case deleted
-
-    /// A string representation of the transfer state.
-    public var label: String {
-        switch self {
-        case .pending:
-            return "Pending"
-        case .inProgress:
-            return "In Progress"
-        case .paused:
-            return "Paused"
-        case .complete:
-            return "Complete"
-        case .failed:
-            return "Failed"
-        case .canceled:
-            return "Canceled"
-        case .deleted:
-            return "Deleted"
-        }
-    }
-
-    /// Indicates whether the transfer is currently in a state that can be resumed.
-    public var resumable: Bool {
-        switch self {
-        case .paused, .failed, .pending, .inProgress:
-            return true
-        case .complete, .canceled, .deleted:
-            return false
-        }
-    }
-
-    /// Indicates whether the transfer is currently in an active state.
-    public var active: Bool {
-        switch self {
-        case .pending, .inProgress:
-            return true
-        case .paused, .failed, .complete, .canceled, .deleted:
-            return false
-        }
-    }
-}
-
-/// The type of a transfer.
-public enum TransferType: Int16 {
-    /// A blob upload.
-    case upload
-    /// A blob download.
-    case download
-
-    /// A string representation of the transfer type.
-    public var label: String {
-        switch self {
-        case .upload:
-            return "upload"
-        case .download:
-            return "download"
-        }
-    }
-}
