+++ sdk/storage/AzureStorageBlob/Source/Operations/MultiBlobOperation.swift
