+++ sdk/storage/AzureStorageBlob/Source/Operations/TransferOperation.swift
