+++ sdk/storage/AzureStorageBlob/Source/Transport/StorageHTTPHeader.swift
